chore: add more Encryption methods
